requirements.txt
pandas
PyMuPDF
pytesseract
Pillow
scikit-learn
joblib
opencv-python-headless
